ci: break up interchaintests into user stories (#3405)
* ci: Set parallelism to 10 for interchaintest

* ci: Move interchaintests into folders based on user stories

* ci: Break up the consumer chain interchaintests into user stories
