+++ tests/interchain/consumer_chain/consumer_launch_test.go
@@ -1,4 +1,4 @@
-package interchain_test
+package consumer_chain_test
 
 import (
 	"testing"
@@ -16,19 +16,19 @@ type ConsumerLaunchSuite struct {
 	OtherChain                   string
 	OtherChainVersionPreUpgrade  string
 	OtherChainVersionPostUpgrade string
-	ShouldCopyProviderKey        [chainsuite.ValidatorCount]bool
+	ShouldCopyProviderKey        []bool
 }
 
-func noProviderKeysCopied() [chainsuite.ValidatorCount]bool {
-	return [chainsuite.ValidatorCount]bool{false, false, false, false, false, false}
+func noProviderKeysCopied() []bool {
+	return []bool{false, false, false, false, false, false}
 }
 
-func allProviderKeysCopied() [chainsuite.ValidatorCount]bool {
-	return [chainsuite.ValidatorCount]bool{true, true, true, true, true, true}
+func allProviderKeysCopied() []bool {
+	return []bool{true, true, true, true, true, true}
 }
 
-func someProviderKeysCopied() [chainsuite.ValidatorCount]bool {
-	return [chainsuite.ValidatorCount]bool{true, false, true, false, true, false}
+func someProviderKeysCopied() []bool {
+	return []bool{true, false, true, false, true, false}
 }
 
 func (s *ConsumerLaunchSuite) TestChainLaunch() {
@@ -93,7 +93,12 @@ func selectConsumerVersion(preV21, postV21 string) string {
 
 func TestICS4ChainLaunch(t *testing.T) {
 	s := &ConsumerLaunchSuite{
-		Suite:                        chainsuite.NewSuite(chainsuite.SuiteConfig{CreateRelayer: true}),
+		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{
+			CreateRelayer: true,
+			ChainSpec: &interchaintest.ChainSpec{
+				NumValidators: &chainsuite.SixValidators,
+			},
+		}),
 		OtherChain:                   "ics-consumer",
 		OtherChainVersionPreUpgrade:  selectConsumerVersion("v4.4.1", "v4.5.0"),
 		OtherChainVersionPostUpgrade: "v4.5.0",
@@ -104,7 +109,12 @@ func TestICS4ChainLaunch(t *testing.T) {
 
 func TestICS6ConsumerAllKeysChainLaunch(t *testing.T) {
 	s := &ConsumerLaunchSuite{
-		Suite:                        chainsuite.NewSuite(chainsuite.SuiteConfig{CreateRelayer: true}),
+		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{
+			CreateRelayer: true,
+			ChainSpec: &interchaintest.ChainSpec{
+				NumValidators: &chainsuite.SixValidators,
+			},
+		}),
 		OtherChain:                   "ics-consumer",
 		OtherChainVersionPreUpgrade:  selectConsumerVersion("v6.0.0", "v6.2.1"),
 		OtherChainVersionPostUpgrade: "v6.2.1",
@@ -115,7 +125,12 @@ func TestICS6ConsumerAllKeysChainLaunch(t *testing.T) {
 
 func TestICS6ConsumerSomeKeysChainLaunch(t *testing.T) {
 	s := &ConsumerLaunchSuite{
-		Suite:                        chainsuite.NewSuite(chainsuite.SuiteConfig{CreateRelayer: true}),
+		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{
+			CreateRelayer: true,
+			ChainSpec: &interchaintest.ChainSpec{
+				NumValidators: &chainsuite.SixValidators,
+			},
+		}),
 		OtherChain:                   "ics-consumer",
 		OtherChainVersionPreUpgrade:  selectConsumerVersion("v6.0.0", "v6.2.1"),
 		OtherChainVersionPostUpgrade: "v6.2.1",
@@ -126,55 +141,16 @@ func TestICS6ConsumerSomeKeysChainLaunch(t *testing.T) {
 
 func TestICS6ConsumerNoKeysChainLaunch(t *testing.T) {
 	s := &ConsumerLaunchSuite{
-		Suite:                        chainsuite.NewSuite(chainsuite.SuiteConfig{CreateRelayer: true}),
+		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{
+			CreateRelayer: true,
+			ChainSpec: &interchaintest.ChainSpec{
+				NumValidators: &chainsuite.SixValidators,
+			},
+		}),
 		OtherChain:                   "ics-consumer",
 		OtherChainVersionPreUpgrade:  selectConsumerVersion("v6.0.0", "v6.2.1"),
 		OtherChainVersionPostUpgrade: "v6.2.1",
 		ShouldCopyProviderKey:        noProviderKeysCopied(),
-	}
-	suite.Run(t, s)
-}
-
-type MainnetConsumerChainsSuite struct {
-	*chainsuite.Suite
-}
-
-func (s *MainnetConsumerChainsSuite) TestMainnetConsumerChainsAfterUpgrade() {
-	// We can't do these consumer launches yet because the chains aren't compatible with launching on v21 yet
-	if semver.Major(s.Env.OldGaiaImageVersion) == s.Env.UpgradeName && s.Env.UpgradeName == "v21" {
-		s.T().Skip("Skipping Consumer Launch tests when going from v21 -> v21")
-	}
-	neutron, err := s.Chain.AddConsumerChain(s.GetContext(), s.Relayer, chainsuite.ConsumerConfig{
-		ChainName:             "neutron",
-		Version:               chainsuite.NeutronVersion,
-		ShouldCopyProviderKey: allProviderKeysCopied(),
-		Denom:                 chainsuite.NeutronDenom,
-		TopN:                  95,
-	})
-	s.Require().NoError(err)
-	stride, err := s.Chain.AddConsumerChain(s.GetContext(), s.Relayer, chainsuite.ConsumerConfig{
-		ChainName:             "stride",
-		Version:               chainsuite.StrideVersion,
-		ShouldCopyProviderKey: allProviderKeysCopied(),
-		Denom:                 chainsuite.StrideDenom,
-		TopN:                  95,
-	})
-	s.Require().NoError(err)
-
-	s.Require().NoError(s.Chain.CheckCCV(s.GetContext(), neutron, s.Relayer, 1_000_000, 0, 1))
-	s.Require().NoError(s.Chain.CheckCCV(s.GetContext(), stride, s.Relayer, 1_000_000, 0, 1))
-
-	s.UpgradeChain()
-
-	s.Require().NoError(s.Chain.CheckCCV(s.GetContext(), neutron, s.Relayer, 1_000_000, 0, 1))
-	s.Require().NoError(s.Chain.CheckCCV(s.GetContext(), stride, s.Relayer, 1_000_000, 0, 1))
-	s.Require().NoError(chainsuite.SendSimpleIBCTx(s.GetContext(), s.Chain, neutron, s.Relayer))
-	s.Require().NoError(chainsuite.SendSimpleIBCTx(s.GetContext(), s.Chain, stride, s.Relayer))
-}
-
-func TestMainnetConsumerChainsAfterUpgrade(t *testing.T) {
-	s := &MainnetConsumerChainsSuite{
-		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{CreateRelayer: true}),
 	}
 	suite.Run(t, s)
 }
