+++ tests/interchain/delegator/testdata/contract.wasm
